pull datastore into separate file
